+++ forge/src/main/java/dev/aika/smsn/forge/mixin/MixinInfo.java
@@ -36,6 +36,7 @@ public class MixinInfo {
                 "dev.aika.smsn.forge.mixin.ribbits.SupportersJSONMixin", "ribbits",
                 "dev.aika.smsn.forge.mixin.supplementaries.CreditsMixin", "supplementaries",
                 "dev.aika.smsn.forge.mixin.titanium.RewardMixin", "titanium",
+                "dev.aika.smsn.forge.mixin.voidscape.DonatorHandlerMixin", "voidscape",
                 "dev.aika.smsn.forge.mixin.xaero.minimap.InternetMixin", "xaerominimap",
                 "dev.aika.smsn.forge.mixin.xaero.minimap.PatreonMixin", "xaerominimap",
                 "dev.aika.smsn.forge.mixin.xaero.world_map.InternetMixin", "xaeroworldmap",