+++ common/src/main/java/dev/aika/smsn/config/SMSNConfigDefault.java
@@ -18,6 +18,7 @@ public class SMSNConfigDefault {
     public static boolean titaniumReward = false;
     public static boolean projecteUUIDCheck = false;
     public static boolean voidscapeDonator = false;
+    public static boolean additionalStructuresSupporterCheck = false;
 
     // QoL
     public static SMSNConfig.QuarkCelebration quarkCelebration = SMSNConfig.QuarkCelebration.HideLGBTQIA;