+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+### Unreleased
+
+**NeoForge**
+
+- Add Mixins for [Hexerei](https://www.curseforge.com/minecraft/mc-mods/hexerei)
+
 ### 1.2.7
 
 **NeoForge**