+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+### Unreleased
+
+**Forge**
+
+- Fewer Quark Celebration
+- Add Mixins for [Voidscape](https://www.curseforge.com/minecraft/mc-mods/voidscape)
+
 ### 1.2.7
 
 **Forge**